requirements.txt
dropbox>=11.36.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.10
python-pptx>=0.6.21